requirements.txt
# This file is intentionally left blank.
pyyaml == 6.0.2